Added code for 771

+++ Algorithms/FifteenTile/PDB15Tile/src/pdb15tile/CalculatePDBHeuristic.java
@@ -26,12 +26,12 @@ public class CalculatePDBHeuristic {
     private byte[][][][][][] six;
     private byte[][][][][][] six2;
 
-    private int[] one;
-    private int[][][][][][][] seven;
-    private int[][][][][][][] seven2;
-    private int[] storedNums1;
-    private int[] storedNums2;
-    private int[] storedNums3;
+    private byte[] one;
+    private byte[][][][][][][] seven;
+    private byte[][][][][][][] seven2;
+    private byte[] storedNums1;
+    private byte[] storedNums2;
+    private byte[] storedNums3;
     public CalculatePDBHeuristic() {
         this.six2 = new byte[16][16][16][16][16][16];
         this.six = new byte[16][16][16][16][16][16];
@@ -41,33 +41,34 @@ public class CalculatePDBHeuristic {
         this.stored2 = new byte[]{7, 8, 11, 12, 14, 15};
         this.stored3 = new byte[]{2, 3, 4};
        
-        this.storedNums1 = new int[]{5};
-        this.storedNums2 = new int[]{1, 2, 3, 4, 6, 7, 8};
-        this.storedNums3 = new int[]{9, 10, 11, 12, 13, 14, 15};
+        this.storedNums1 = new byte[]{5};
+        this.storedNums2 = new byte[]{1, 2, 3, 4, 6, 7, 8};
+        this.storedNums3 = new byte[]{9, 10, 11, 12, 13, 14, 15};
 
     }
 
     public void getStoredFiles() throws IOException, ClassNotFoundException {
-        
+        /*
         ObjectInputStream inputStream = new ObjectInputStream(new FileInputStream("PDB_3_1"));
         ObjectInputStream inputStream1 = new ObjectInputStream(new FileInputStream("PDB_6_1"));
         ObjectInputStream inputStream2 = new ObjectInputStream(new FileInputStream("PDB_6_2"));
         three = (byte[][][]) inputStream.readObject();
         six = (byte[][][][][][]) inputStream1.readObject();
         six2 = (byte[][][][][][]) inputStream2.readObject();
-         /*
+        */
+        
         ObjectInputStream inputStream = new ObjectInputStream(new FileInputStream("PDB_1_1"));
         ObjectInputStream inputStream1 = new ObjectInputStream(new FileInputStream("PDB_7_1"));
         ObjectInputStream inputStream2 = new ObjectInputStream(new FileInputStream("PDB_7_2"));
-        one = (int[]) inputStream.readObject();
-        seven = (int[][][][][][][]) inputStream1.readObject();
-        seven2 = (int[][][][][][][]) inputStream2.readObject();
-        */
+        one = (byte[]) inputStream.readObject();
+        seven = (byte[][][][][][][]) inputStream1.readObject();
+        seven2 = (byte[][][][][][][]) inputStream2.readObject();
+        
     }
 
     public byte calculate663(byte[] currentState) {
         byte total = 0;
-        byte[] temp = new byte[3];
+        int[] temp = new int[3];
         for (byte i = 0; i < stored3.length; i++) {
             byte value = getPatternPosition(stored3[i], currentState);
             temp[i] = value;
@@ -89,7 +90,7 @@ public class CalculatePDBHeuristic {
         return total;
     }
 
-    /*
+   
     public byte calculate771(byte[] currentState) {
         byte total = 0;
         int[] temp = new int[1];
@@ -104,8 +105,6 @@ public class CalculatePDBHeuristic {
             int value = getPatternPosition(storedNums2[i], currentState);
             temp2[i] = value;
         }
-        // System.out.println(" H: " + seven[temp2[0]][temp2[1]][temp2[2]][temp2[3]][temp2[4]][temp2[5]][temp2[6]]);
-        
         total += seven[temp2[0]][temp2[1]][temp2[2]][temp2[3]][temp2[4]][temp2[5]][temp2[6]];
 
         int[] temp3 = new int[7];
@@ -116,7 +115,7 @@ public class CalculatePDBHeuristic {
         total += seven2[temp3[0]][temp3[1]][temp3[2]][temp3[3]][temp3[4]][temp3[5]][temp3[6]];
         return total;
     }
-    */
+   
     public byte getPatternPosition(byte value, byte[] tilesInPattern) {
         for (byte i = 0; i < 16; i++) {
             byte tile = tilesInPattern[i];
@@ -127,7 +126,7 @@ public class CalculatePDBHeuristic {
         return -1; //not found
     }
 
-    public int calculateManhattan(int[] puzz) {
+    public int calculateManhattan(byte[] puzz) {
         int total = 0;
         for (int j = 0; j < puzz.length; j++) {
             int i = puzz[j];
